profile image cache removed and renamed
